pyproject.toml
[project]
name = "yolo-label-detector"
version = "0.1.0"
description = ""
authors = [
    {name = "Alena Bannikova",email = "145114139+lekyshka@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ultralytics (>=8.3.235,<9.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "mlflow (>=3.7.0,<4.0.0)",
    "dvc[gdrive] (>=3.64.2,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.9.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pre-commit (>=4.5.0,<5.0.0)"
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
